🚨 Fix linter warnings

+++ micras_nav/src/grid_pose.cpp
@@ -12,7 +12,7 @@
 
 namespace micras::nav {
 Side angle_to_grid(float angle) {
-    int8_t grid_angle = std::lround(2.0F * angle / std::numbers::pi_v<float>);
+    int32_t grid_angle = std::lround(2.0F * angle / std::numbers::pi_v<float>);
     return static_cast<Side>(grid_angle < 0 ? 4 + (grid_angle % 4) : grid_angle % 4);
 }
